refactor: rename storage.service.ts to in-memory-db.service.ts
