+++ src/shared/shared.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { InMemoryDbService } from './services/storage.service';
+import { InMemoryDbService } from './services/in-memory-db.service';
 import { UUIDService } from './services/uuid.service';
 
 @Module({
